Update _code-of-conduct.pl.md
I find the word "usunięcie" (also "wydalenie") a better one than "wyrzucenie" for translation of "expulsion". "Wyrzucenie" seems to be something stronger and less formal, not fitting that good in the code of conduct. Of course that is only my opinion, I'm happy to hear how others think of such proposed change. Do you perceive it similarly?